pyproject.toml
[project]
name = "Cloudberry-Storage-Any-Modal"
version = "0.1.0"
description = "gRPC multimodal storage over Qdrant"
authors = [{ name = "sckwokyboom", email = "sckwokyboom@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
  "grpcio",
  "grpcio-tools",
  "protobuf",
  "numpy",
  "pillow",
  "qdrant-client",
  "deep-translator",
  "pytest",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[project.scripts]
generate-proto = "scripts.generate_proto:main"